atualizei

+++ src/main/java/com/belchiorsapalo/codeFormater/problem/DTO/ProblemRegisterDTO.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 import com.belchiorsapalo.codeFormater.testCase.model.TestCase;
 
-public record ProblemRegisterDTO(String title, String description, int sequence, double points, List<TestCase> testCases) {
+public record ProblemRegisterDTO(String title, String description, int sequence, int points, List<TestCase> testCases) {
 }
